// packages/ui-tasks/src/boards/types.ts
export interface IOptions {
  EditForm: any;
  Item: any;
  type: string;
  title: string;
  queriesName: {
    itemsQuery: string;
    itemsTotalCountQuery: string;
    detailQuery: string;
    archivedItemsQuery: string;
    archivedItemsCountQuery: string;
  };
  mutationsName: {
    addMutation: string;
    editMutation: string;
    removeMutation: string;
    changeMutation: string;
    watchMutation: string;
    copyMutation: string;
    archiveMutation: string;
  };
  queries: {
    itemsQuery: string;
    itemsTotalCountQuery: string;
    detailQuery: string;
    archivedItemsQuery: string;
    archivedItemsCountQuery: string;
  };
  mutations: {
    addMutation: string;
    editMutation: string;
    removeMutation: string;
    changeMutation: string;
    watchMutation: string;
    archiveMutation: string;
    copyMutation: string;
  };
  texts: {
    addText: string;
    addSuccessText?: string;
    updateSuccessText: string;
    deleteSuccessText: string;
    copySuccessText: string;
    changeSuccessText: string;
  };
  isMove: boolean;
  getExtraParams: (queryParams: any) => any;
}

export interface IPipeline {
  _id: string;
  name: string;
  boardId: string;
  tagId?: string;
  visibility: string;
  status: string;
  createdAt: Date;
  createdUser: IUser;
  members?: any[];
  departmentIds?: string[];
  memberIds?: string[];
  condition?: string;
  label?: string;
  bgColor?: string;
  isWatched: boolean;
  startDate?: Date;
  endDate?: Date;
  metric?: string;
  hackScoringType?: string;
  templateId?: string;
  state?: string;
  itemsTotalCount?: number;
  isCheckDate?: boolean;
  isCheckUser?: boolean;
  isCheckDepartment?: boolean;
  excludeCheckUserIds?: string[];
  numberConfig?: string;
  numberSize?: string;
  nameConfig?: string;
}

export interface IBoard {
  _id: string;
  name: string;
  pipelines?: IPipeline[];
}

export interface IItemParams {
  _id?: string;
  name?: string;
  stageId?: string;
  parentId?: string;
  assignedUserIds?: string[];
  closeDate?: Date;
  description?: string;
  order?: number;
  isComplete?: boolean;
  priority?: string;
  reminderMinute?: number;
  companyIds?: string[];
  customerIds?: string[];
  sourceConversationIds?: string[];
  labelIds?: string[];
  proccessId?: string;
  aboveItemId?: string;
  attachments?: string[];
  relationData?: any;
  branchIds?: string[];
  departmentIds?: string[];
}

export type SaveItemMutation = ({ variables: IItemParams }) => Promise<any>;
export type RemoveStageMutation = ({
  variables,
}: {
  variables: { _id: string };
}) => Promise<any>;

interface IStageComparisonInfo {
  count: number;
  percent: number;
}

export interface IStage {
  _id: string;
  name: string;
  type: string;
  probability: string;
  index?: number;
  itemId?: string;
  unUsedAmount?: any;
  amount?: any;
  itemsTotalCount: number;
  formId: string;
  pipelineId: string;
  visibility: string;
  memberIds: string[];
  canMoveMemberIds?: string[];
  canEditMemberIds?: string[];
  departmentIds: string[];
  status: string;
  order: number;
  code?: string;
  age?: number;
  defaultTick?: boolean;
}
// ...
export interface IPipelineLabel {
  _id?: string;
  name: string;
  colorCode: string;
  pipelineId?: string;
  createdBy?: string;
  createdAt?: Date;
}

export interface IPipelineLabelVariables {
  name: string;
  colorCode: string;
  pipelineId: string;
}
export interface IItem {
  _id: string;
  name: string;
  order: number;
  stageId: string;
  boardId?: string;
  startDate: Date;
  closeDate: Date;
  description: string;
  unUsedAmount?: number;
  amount: number;
  modifiedAt: Date;
  assignedUserIds?: string[];
  assignedUsers: IUser[];
  createdUser?: IUser;
  companies: ICompany[];
  customers: ICustomer[];
  attachments?: IAttachment[];
  labels: IPipelineLabel[];
  pipeline: IPipeline;
  stage?: IStage;
  isWatched?: boolean;
  priority?: string;
  hasNotified?: boolean;
  isComplete: boolean;
  reminderMinute: number;
  labelIds: string[];
  status?: string;
  createdAt: Date;
  timeTrack: {
    status: string;
    timeSpent: number;
    startDate?: string;
  };
  customFieldsData?: {
    [key: string]: any;
  };
  score?: number;
  number?: string;
  relations: any[];
  tags: ITag[];
  tagIds: string[];
  customProperties?: any;
  departmentIds: string[];
  branchIds: string[];
}
// ...
export interface IEditFormContent {
  state: any;
  saveItem: (doc: { [key: string]: any }, callback?: (item) => void) => void;
  onChangeStage: (stageId: string) => void;
  copy: () => void;
  remove: (id: string) => void;
}
// ...
export type Item = {
  name: string;
  color: string;
};